Remove Legacy Works With Nest
