--- homeassistant/components/nest/legacy/const.py
@@ -1,6 +1,0 @@
-"""Constants used by the legacy Nest component."""
-
-DOMAIN = "nest"
-DATA_NEST = "nest"
-DATA_NEST_CONFIG = "nest_config"
-SIGNAL_NEST_UPDATE = "nest_update"
