+++ tests/components/nest/test_camera.py
