+++ tests/components/nest/test_diagnostics.py
@@ -9,8 +9,6 @@ from homeassistant.config_entries import ConfigEntryState
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import device_registry as dr
 
-from .common import TEST_CONFIG_LEGACY
-
 from tests.components.diagnostics import (
     get_diagnostics_for_config_entry,
     get_diagnostics_for_device,
@@ -142,21 +140,6 @@ async def test_setup_susbcriber_failure(
         await setup_base_platform()
 
     assert config_entry.state is ConfigEntryState.SETUP_RETRY
-
-    assert await get_diagnostics_for_config_entry(hass, hass_client, config_entry) == {}
-
-
-@pytest.mark.parametrize("nest_test_config", [TEST_CONFIG_LEGACY])
-async def test_legacy_config_entry_diagnostics(
-    hass: HomeAssistant,
-    hass_client: ClientSessionGenerator,
-    config_entry,
-    setup_base_platform,
-) -> None:
-    """Test config entry diagnostics for legacy integration doesn't fail."""
-
-    with patch("homeassistant.components.nest.legacy.Nest"):
-        await setup_base_platform()
 
     assert await get_diagnostics_for_config_entry(hass, hass_client, config_entry) == {}
